v0.04 : Done	:	Toggle for Tool & Idol. Income with tool and building bonus (still missing upgrade, synergy & ascension).

+++ src/modele/intel/Intel.java
@@ -81,8 +81,8 @@ public class Intel extends MilkXmlObj implements Cloneable {
 		this.start = start;
 	}
 	
-	public Float getPriceValue() {
-		return this.price.getCoin()*this.price.getCoef();
+	public Double getPriceValue() {
+		return (double) (this.price.getCoin()*this.price.getCoef());
 	}
 	public Price getPrice() {
 		return this.price;
